+++ packages/disclosure-button/src/types.ts
@@ -1,0 +1,5 @@
+export interface DisclosureButtonDetail {
+  ids: string[];
+  elements: HTMLElement[];
+  el: HTMLElement;
+}